requirements.txt
selenium==4.23.0
beautifulsoup4==4.12.3
scikit-learn==1.5.1
scipy==1.14.0
pandas==2.2.2
tensorflow==2.15.0
joblib==1.4.2
numpy==1.23.5
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.7
openpyxl==3.1.5
pandas==2.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
